Remove hardcoded -m32 in compiler options
It's required for x86_64 targets where we obviously don't want to
enforce -m32.
It shouldn't affect standard i686 targets because i686 is enforced on
configure stage and it effectively removes 64-bit support, keeping -m32
default.
Patch changes 4.7 and 4.8 only. x86_64 targets are not expected to be
supported in older gcc versions due to many issues and dependencies.

Change-Id: Id15a677a9d3b1712eb95aba92c2c5bd98f9ed1cc
Signed-off-by: Pavel Chupin <pavel.v.chupin@intel.com>